pyproject.toml
[project]
name = "agentception"
version = "0.1.1"
description = "Multi-agent orchestration system for AI-powered development workflows"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "markdown>=3.7",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.4",
    "aiofiles>=23.2.1",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "sse-starlette>=2.1.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "anyio>=4.0.0",
    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-anyio>=0.0.0",
    "pytest-timeout>=2.0.0",
    "mypy>=1.8.0",
    "types-aiofiles>=23.2.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
agentception = "agentception.app:app"

[project.urls]
Repository = "https://github.com/cgcardona/agentception"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentception*"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
explicit_package_bases = true
warn_unreachable = true
show_error_codes = true

# Third-party stubs are not always available
[[tool.mypy.overrides]]
module = ["sqlalchemy.*", "alembic.*", "asyncpg.*", "aiosqlite.*", "markdown"]
ignore_missing_imports = true

# resolve_arch lives in scripts/gen_prompts/ and is not a package — added to
# sys.path at runtime.  Suppress the import-not-found error here rather than
# using a # type: ignore at every call site.
[[tool.mypy.overrides]]
module = ["resolve_arch"]
ignore_missing_imports = true

# pytest decorators are untyped by design
[[tool.mypy.overrides]]
module = ["agentception.tests.*"]
disallow_untyped_decorators = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["agentception/tests"]
timeout = 30
filterwarnings = [
    "ignore::pytest.PytestUnraisableExceptionWarning",
]
markers = [
    "e2e: end-to-end smoke tests that exercise the full workflow with mocked GitHub",
]

[tool.black]
line-length = 100
target-version = ["py312"]
